requirements.txt
asgiref==3.7.2
Django==4.2.2
djangorestframework==3.14.0
environs==9.5.0
marshmallow==3.19.0
packaging==23.1
psycopg==3.1.9
psycopg-binary==3.1.9
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
